+++ src/ts/src/public/system-config/SystemConfigGetRequest.ts
@@ -1,0 +1,5 @@
+import { SystemConfigDto } from "./dto";
+
+export interface SystemConfigGetResponse {
+  item: SystemConfigDto;
+}