+++ src/ts/src/public/system-config/api.ts
@@ -1,0 +1,11 @@
+import { IApiClient } from "../IApiClient";
+import { SystemConfigGetResponse } from "./SystemConfigGetRequest";
+
+export class SystemConfigApi {
+  constructor(private readonly http: IApiClient) { }
+  #prefix = '/api/v1/system-config';
+
+  get() {
+    return this.http.getJson<SystemConfigGetResponse>(`${this.#prefix}`);
+  }
+}