+++ src/ts/src/public/system-config/events.ts
@@ -1,0 +1,6 @@
+import type { SystemConfigDto } from "./dto";
+
+
+export interface SystemConfigSetEvent {
+  item: SystemConfigDto;
+}