+++ src/ts/src/public/system-config/index.ts
@@ -1,0 +1,4 @@
+export * from "./api";
+export * from "./dto";
+export * from "./events";
+export * from "./SystemConfigGetRequest";