+++ src/ts/src/public/workspace/values.ts
@@ -1,4 +1,3 @@
 export type WorkspaceMemberRole = "Owner" | "Member";
-export type WorkspacePlan = "FreeWorkspace" | "UpgradedWorkspace" | "ExtraWorkspace" | "TutorialTemporaryWorkspace" | "TutorialTemplateWorkspace";
-export type WorkspaceType = "Free" | "Paid" | "Special";
-export type WorkspaceStatus = "Provisioning" | "Active" | "Deleted" | "Paused";
+export type WorkspaceType = "Free" | "Paid" | "Temp";
+export type WorkspaceStatus = "Provisioning" | "Active" | "Deleted";